+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
+### Unreleased
+### Added
+- Add two configuration options to include (`includePatterns`) and exclude (`excludePatterns`) files to be uploaded. See [configuration documentation](.docs/configuration.md) for details.
+
+### Fixed
+- Fix folder typo on secrets backend storage
+
 ## 0.3.2
 ### Added
 - Add `go get` installation method to [README](README.md)